TinyInvoice 7.3: Fix exception
When removing a role via /administrative/role, I ran into the following exception:

Argument 1 passed to Document_Tag::get_by_tag() must be an instance of Tag, instance of Role given, called in /var/www/bookkeeping.test.tickoweb.be/lib/model/Role.php on line 26

This seemed to be due to a call in Role to Document_Tag::get_by_tag($this). Since document tags have nothing to do with roles, and as such document_tags do not need to
be removed together with the role. So this loop has now been removed. The only line left in the delete method was $this->trait_delete, so I removed the
delete method altogether and updated the delete trait use statement.